Creates a Comparison Queue to help with writing the Serial Interface
The comparison queue takes in bytes one at a time. Each byte taken in
is compared to the expected phrase which starts command mode: $ESP32$
If a byte coming in does not match a character, the function returns
the number of bytes that should be flushed from the buffer and the
buffer address. In this way, data can be read to search for the presence
of the command phrase, or otherwise simply forwarded.